Django-Advance-Blog/requirements.txt

# general modules
django >3.2,<3.3
python-decouple
Pillow
djangorestframework

# Third party modules
Markdown
django-filter
drf-yasg[validation]
djangorestframework-simplejwt


# Deployment modules
